requirements.txt
streamlit
torch
torchvision
Pillow
albumentations
numpy
twilio